refactor: split tunnel helpers
