+++ src/backend/ssh/tunnel.ts
@@ -16,7 +16,6 @@ import type {
   TunnelConfig,
   TunnelStatus,
   VerificationData,
-  ErrorType,
   AuthenticatedRequest,
 } from "../../types/index.js";
 import { CONNECTION_STATES } from "../../types/index.js";
@@ -28,6 +27,21 @@ import { createSocks5Connection } from "../utils/socks5-helper.js";
 import { AuthManager } from "../utils/auth-manager.js";
 import { PermissionManager } from "../utils/permission-manager.js";
 import { withConnection } from "./ssh-connection-pool.js";
+import {
+  applyAuthOptions,
+  bindForwardIn,
+  connectClient,
+  forwardOut,
+  getManagedTunnelAlgorithms,
+  pipeTunnelStreams,
+  unbindForwardIn,
+} from "./tunnel-ssh-primitives.js";
+import {
+  classifyTunnelError,
+  getTunnelMarker,
+  normalizeTunnelName,
+  validateTunnelConfig,
+} from "./tunnel-utils.js";
 
 const app = express();
 app.use(createCorsMiddleware(["GET", "POST", "PUT", "DELETE", "OPTIONS"]));
@@ -54,7 +68,7 @@ const retryExhaustedTunnels = new Set<string>();
 const cleanupInProgress = new Set<string>();
 const tunnelConnecting = new Set<string>();
 const lastTunnelErrors = new Map<string, string>();
-const lastTunnelErrorTypes = new Map<string, ErrorType>();
+const lastTunnelErrorTypes = new Map<string, TunnelStatus["errorType"]>();
 
 const tunnelConfigs = new Map<string, TunnelConfig>();
 const activeTunnelProcesses = new Map<string, ChildProcess>();
@@ -201,121 +215,6 @@ function broadcastTunnelStatusSnapshot(): void {
   for (const client of tunnelStatusClients) {
     sendTunnelStatusSnapshot(client);
   }
-}
-
-function classifyError(errorMessage: string): ErrorType {
-  if (!errorMessage) return "UNKNOWN";
-
-  const message = errorMessage.toLowerCase();
-
-  if (
-    message.includes("closed by remote host") ||
-    message.includes("connection reset by peer") ||
-    message.includes("connection refused") ||
-    message.includes("broken pipe")
-  ) {
-    return "NETWORK_ERROR";
-  }
-
-  if (
-    message.includes("authentication failed") ||
-    message.includes("permission denied") ||
-    message.includes("incorrect password")
-  ) {
-    return "AUTHENTICATION_FAILED";
-  }
-
-  if (
-    message.includes("connect etimedout") ||
-    message.includes("timeout") ||
-    message.includes("timed out") ||
-    message.includes("keepalive timeout")
-  ) {
-    return "TIMEOUT";
-  }
-
-  if (
-    message.includes("bind: address already in use") ||
-    message.includes("failed for listen port") ||
-    message.includes("port forwarding failed")
-  ) {
-    return "CONNECTION_FAILED";
-  }
-
-  if (message.includes("permission") || message.includes("access denied")) {
-    return "CONNECTION_FAILED";
-  }
-
-  return "UNKNOWN";
-}
-
-function getTunnelMarker(tunnelName: string) {
-  return `TUNNEL_MARKER_${tunnelName.replace(/[^a-zA-Z0-9]/g, "_")}`;
-}
-
-function normalizeTunnelName(
-  hostId: number,
-  tunnelIndex: number,
-  displayName: string,
-  sourcePort: number,
-  endpointHost: string,
-  endpointPort: number,
-): string {
-  return `${hostId}::${tunnelIndex}::${displayName}::${sourcePort}::${endpointHost}::${endpointPort}`;
-}
-
-function parseTunnelName(tunnelName: string): {
-  hostId?: number;
-  tunnelIndex?: number;
-  displayName: string;
-  sourcePort: string;
-  endpointHost: string;
-  endpointPort: string;
-  isLegacyFormat: boolean;
-} {
-  const parts = tunnelName.split("::");
-
-  if (parts.length === 6) {
-    return {
-      hostId: parseInt(parts[0]),
-      tunnelIndex: parseInt(parts[1]),
-      displayName: parts[2],
-      sourcePort: parts[3],
-      endpointHost: parts[4],
-      endpointPort: parts[5],
-      isLegacyFormat: false,
-    };
-  }
-
-  tunnelLogger.warn(`Legacy tunnel name format: ${tunnelName}`);
-
-  const legacyParts = tunnelName.split("_");
-  return {
-    displayName: legacyParts[0] || "unknown",
-    sourcePort: legacyParts[legacyParts.length - 3] || "0",
-    endpointHost: legacyParts[legacyParts.length - 2] || "unknown",
-    endpointPort: legacyParts[legacyParts.length - 1] || "0",
-    isLegacyFormat: true,
-  };
-}
-
-function validateTunnelConfig(
-  tunnelName: string,
-  tunnelConfig: TunnelConfig,
-): boolean {
-  const parsed = parseTunnelName(tunnelName);
-
-  if (parsed.isLegacyFormat) {
-    return true;
-  }
-
-  return (
-    parsed.hostId === tunnelConfig.sourceHostId &&
-    parsed.tunnelIndex === tunnelConfig.tunnelIndex &&
-    String(parsed.sourcePort) === String(tunnelConfig.sourcePort) &&
-    parsed.endpointHost === tunnelConfig.endpointHost &&
-    String(parsed.endpointPort) === String(tunnelConfig.endpointPort)
-  );
 }
 
 async function cleanupTunnelResources(
@@ -639,183 +538,6 @@ function getTunnelScope(tunnelConfig: TunnelConfig): "s2s" | "c2s" {
 
 function getTunnelBindHost(tunnelConfig: TunnelConfig): string {
   return tunnelConfig.bindHost || "127.0.0.1";
-}
-
-function getManagedTunnelAlgorithms() {
-  return {
-    kex: [
-      "curve25519-sha256",
-      "curve25519-sha256@libssh.org",
-      "ecdh-sha2-nistp521",
-      "ecdh-sha2-nistp384",
-      "ecdh-sha2-nistp256",
-      "diffie-hellman-group-exchange-sha256",
-      "diffie-hellman-group14-sha256",
-      "diffie-hellman-group14-sha1",
-      "diffie-hellman-group-exchange-sha1",
-      "diffie-hellman-group1-sha1",
-    ],
-    serverHostKey: [
-      "ssh-ed25519",
-      "ecdsa-sha2-nistp521",
-      "ecdsa-sha2-nistp384",
-      "ecdsa-sha2-nistp256",
-      "rsa-sha2-512",
-      "rsa-sha2-256",
-      "ssh-rsa",
-      "ssh-dss",
-    ],
-    cipher: SSH_ALGORITHMS.cipher,
-    hmac: [
-      "hmac-sha2-512-etm@openssh.com",
-      "hmac-sha2-256-etm@openssh.com",
-      "hmac-sha2-512",
-      "hmac-sha2-256",
-      "hmac-sha1",
-      "hmac-md5",
-    ],
-    compress: ["none", "zlib@openssh.com", "zlib"],
-  };
-}
-
-function applyAuthOptions(
-  connOptions: Record<string, unknown>,
-  credentials: {
-    password?: string;
-    sshKey?: string;
-    keyPassword?: string;
-    keyType?: string;
-    authMethod?: string;
-  },
-): void {
-  if (credentials.authMethod === "key" && credentials.sshKey) {
-    const cleanKey = credentials.sshKey
-      .trim()
-      .replace(/\r\n/g, "\n")
-      .replace(/\r/g, "\n");
-    connOptions.privateKey = Buffer.from(cleanKey, "utf8");
-    if (credentials.keyPassword) {
-      connOptions.passphrase = credentials.keyPassword;
-    }
-    if (credentials.keyType && credentials.keyType !== "auto") {
-      connOptions.privateKeyType = credentials.keyType;
-    }
-  } else {
-    connOptions.password = credentials.password;
-  }
-}
-
-function connectClient(
-  connOptions: Record<string, unknown>,
-  tunnelName: string,
-  role: "source" | "endpoint",
-): Promise<Client> {
-  return new Promise((resolve, reject) => {
-    const client = new Client();
-    let settled = false;
-    client.once("ready", () => {
-      settled = true;
-      resolve(client);
-    });
-    client.once("error", (error) => {
-      if (!settled) {
-        reject(error);
-        return;
-      }
-      tunnelLogger.error("Managed tunnel SSH client error", error, {
-        operation: "managed_tunnel_client_error",
-        tunnelName,
-        role,
-      });
-    });
-    client.connect(connOptions);
-  });
-}
-
-function forwardOut(
-  client: Client,
-  targetHost: string,
-  targetPort: number,
-  tunnelName?: string,
-): Promise<ClientChannel> {
-  return new Promise((resolve, reject) => {
-    client.forwardOut("127.0.0.1", 0, targetHost, targetPort, (err, stream) => {
-      if (err) {
-        if (tunnelName) {
-          tunnelLogger.error("Managed tunnel forwardOut failed", err, {
-            operation: "managed_tunnel_forward_out_failed",
-            tunnelName,
-            targetHost,
-            targetPort,
-          });
-        }
-        reject(err);
-        return;
-      }
-      resolve(stream);
-    });
-  });
-}
-
-function bindForwardIn(
-  client: Client,
-  bindHost: string,
-  bindPort: number,
-): Promise<number> {
-  return new Promise((resolve, reject) => {
-    client.forwardIn(bindHost, bindPort, (err, actualPort) => {
-      if (err) {
-        reject(err);
-        return;
-      }
-      resolve(actualPort || bindPort);
-    });
-  });
-}
-
-function unbindForwardIn(
-  client: Client,
-  bindHost: string,
-  bindPort: number,
-): void {
-  try {
-    client.unforwardIn(bindHost, bindPort, (err) => {
-      if (err) {
-        tunnelLogger.warn("Failed to unbind managed tunnel listener", {
-          operation: "managed_tunnel_unforward_failed",
-          bindHost,
-          bindPort,
-          error: err.message,
-        });
-      }
-    });
-  } catch {
-    // The connection may already be gone.
-  }
-}
-
-function pipeTunnelStreams(
-  inbound: Duplex,
-  outboundPromise: Promise<Duplex>,
-  tunnelName: string,
-): void {
-  outboundPromise
-    .then((outbound) => {
-      inbound.pipe(outbound).pipe(inbound);
-      inbound.on("error", () => outbound.destroy());
-      outbound.on("error", () => inbound.destroy());
-    })
-    .catch((error) => {
-      tunnelLogger.error(
-        "Failed to open managed tunnel outbound stream",
-        error,
-        {
-          operation: "managed_tunnel_outbound_failed",
-          tunnelName,
-        },
-      );
-      inbound.destroy();
-    });
 }
 
 function parseSocksAddress(buffer: Buffer): {
@@ -1805,7 +1527,7 @@ async function connectSSHTunnel(
   conn.on("error", (err) => {
     clearTimeout(connectionTimeout);
 
-    const errorType = classifyError(err.message);
+    const errorType = classifyTunnelError(err.message);
 
     tunnelLogger.error(`Tunnel connection failed for '${tunnelName}'`, err, {
       operation: "tunnel_connect_error",
@@ -1922,7 +1644,7 @@ async function connectSSHTunnel(
     } catch (error) {
       const message =
         error instanceof Error ? error.message : "Failed to create tunnel";
-      const errorType = classifyError(message);
+      const errorType = classifyTunnelError(message);
       tunnelLogger.error("Failed to create managed tunnel", error, {
         operation: "managed_tunnel_create_failed",
         tunnelName,
